requirements.txt
absl-py==1.0.0
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.2.2
#anaconda-navigator
#anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
#asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.3.0
attrs==19.3.0
#Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
autopep8==1.4.4
autoscraper==1.1.12
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
blaze==0.10.1
bleach==3.1.0
blinker==1.4
blis==0.7.6
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.2
Brotli==1.0.9
bs4==0.0.1
cachetools==4.1.0
catalogue==2.0.6
cchardet==2.1.7
certifi==2021.10.8
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.3.0
clyent==1.2.1
colorama==0.4.3
colorlover==0.3.0
comtypes==1.1.7
conda==4.2.7
conda-build==2.1.5
#conda-package-handling==1.6.0
#conda-verify==3.4.2
configparser==5.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
crayons==0.4.0
cryptography==2.8
cssselect==1.1.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.6
Cython==0.29.15
cytoolz==0.10.1
dash==1.21.0
dash-bootstrap-components==0.12.2
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
dask==2.11.0
dataclasses==0.6
datashape==0.5.2
dateparser==1.1.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.11.0
Django==3.1.2
dnspython==1.16.0
docutils==0.16
elastic-transport==8.0.1
elasticsearch==8.0.1
#en-core-web-sm==3.2.0
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
file-operations==1.0.5
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Compress==1.10.1
#Flask-Cors @ file:///tmp/build/80754af9/flask-cors_1609957731919/work
frozenlist==1.3.0
fsspec==0.6.2
future==0.18.2
gast==0.3.3
#gdelt==0.1.10.6.1
geographiclib==1.52
geopy==2.2.0
gevent==1.4.0
gitdb==4.0.7
GitPython==3.1.18
glob2==0.7
google-auth==1.17.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletransx==2.4.2
greenlet==0.4.15
grpcio==1.29.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
#hyperlink @ file:///tmp/build/80754af9/hyperlink_1610130746837/work
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imagesize==1.2.0
#imbalanced-learn @ file:///home/conda/feedstock_root/build_artifacts/imbalanced-learn_1591716883293/work
importlib-metadata==1.5.0
#incremental @ file:///tmp/build/80754af9/incremental_1636629750599/work
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jieba3k==0.35.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
#jupyterlab-launcher @ file:///tmp/build/80754af9/jupyterlab_launcher_1601490670252/work
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.1.0
kiwisolver==1.1.0
kneed==0.7.0
langcodes==3.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
livelossplot==0.5.1
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
#mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==6.0.2
multipledispatch==0.6.0
murmurhash==1.0.6
mysql-connector-python==8.0.20
#navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.4
networkx==2.4
newspaper3k==0.2.8
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
#numpy==1.21.5
numpydoc==0.9.2
oauthlib==3.2.0
odo==0.5.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.1
#pandas==1.0.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
pathy==0.6.1
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plotly==4.9.0
pluggy==0.13.1
ply==3.11
preshed==3.0.6
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.12.2
psutil==5.6.7
#psycopg2 @ file:///C:/ci/psycopg2_1612298715048/work
psycopg2-binary==2.9.1
py==1.8.1
pyarrow==4.0.1
#pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycares==4.1.2
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
#pycurl==7.43.0.5
pydantic==1.8.2
pydeck==0.6.2
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
#PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1615748656804/work
pylint==2.4.4
pymongo==3.11.0
PyNaCl==1.3.0
pyodbc===4.0.32
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-socks==2.0.3
pyttsx3==2.90
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex==2021.11.10
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.2
#Rtree==0.9.3
ruamel_yaml==0.15.87
schedule==1.1.0
scikit-image==0.16.2
#scikit-learn @ file:///C:/ci/scikit-learn_1642599122269/work
scikit-plot==0.3.7
#scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
#service-identity @ file:///Users/ktietz/demo/mc3/conda-bld/service_identity_1629460757137/work
sgmllib3k==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==2.4.2
#statsmodels==0.11.0
streamlit==0.85.1
sympy==1.5.1
#tables==3.6.1
tblib==1.6.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.8.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==8.0.13
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.2.0
toml==0.10.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
tweepy==4.6.0
tweet-preprocessor==0.6.0
-e git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221#egg=twint
#Twisted @ file:///C:/ci/twisted_1636996748528/work
#twisted-iocpsupport @ file:///C:/ci/twisted-iocpsupport_1614804836190/work
typer==0.4.0
#typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.26.8
validators==0.18.2
wasabi==0.9.0
watchdog==0.10.2
wcwidth==0.1.8
webdriver-manager==3.4.2
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xgboost==1.4.2
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
yarl==1.7.2
zict==1.0.0
zipp==2.2.0
#zope.interface @ file:///C:/ci/zope.interface_1625036257680/work
